+++ griptape/engines/rag/modules/response/prompt_response_rag_module.py
@@ -1,27 +1,30 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING, Any, Callable, Optional
 
 from attrs import Factory, define, field
 
 from griptape.artifacts.text_artifact import TextArtifact
 from griptape.engines.rag.modules import BaseResponseRagModule
+from griptape.mixins import RuleMixin
 from griptape.utils import J2
 
 if TYPE_CHECKING:
+    from griptape.artifacts import BaseArtifact
     from griptape.drivers import BasePromptDriver
     from griptape.engines.rag import RagContext
 
 
 @define(kw_only=True)
-class PromptResponseRagModule(BaseResponseRagModule):
-    answer_token_offset: int = field(default=400)
+class PromptResponseRagModule(BaseResponseRagModule, RuleMixin):
     prompt_driver: BasePromptDriver = field()
+    answer_token_offset: int = field(default=400)
+    metadata: Optional[str] = field(default=None)
     generate_system_template: Callable[[RagContext, list[TextArtifact]], str] = field(
         default=Factory(lambda self: self.default_system_template_generator, takes_self=True),
     )
 
-    def run(self, context: RagContext) -> RagContext:
+    def run(self, context: RagContext) -> BaseArtifact:
         query = context.query
         tokenizer = self.prompt_driver.tokenizer
         included_chunks = []
@@ -45,15 +48,17 @@ class PromptResponseRagModule(BaseResponseRagModule):
         output = self.prompt_driver.run(self.generate_prompt_stack(system_prompt, query)).to_artifact()
 
         if isinstance(output, TextArtifact):
-            context.output = output
+            return output
         else:
             raise ValueError("Prompt driver did not return a TextArtifact")
 
-        return context
-
     def default_system_template_generator(self, context: RagContext, artifacts: list[TextArtifact]) -> str:
-        return J2("engines/rag/modules/response/prompt/system.j2").render(
-            text_chunks=[c.to_text() for c in artifacts],
-            before_system_prompt="\n\n".join(context.before_query),
-            after_system_prompt="\n\n".join(context.after_query),
-        )
+        params: dict[str, Any] = {"text_chunks": [c.to_text() for c in artifacts]}
+
+        if len(self.all_rulesets) > 0:
+            params["rulesets"] = J2("rulesets/rulesets.j2").render(rulesets=self.all_rulesets)
+
+        if self.metadata is not None:
+            params["metadata"] = J2("engines/rag/modules/response/metadata/system.j2").render(metadata=self.metadata)
+
+        return J2("engines/rag/modules/response/prompt/system.j2").render(**params)